moving ./pruning to ./pruner/pruning setting the default pruning in training.py to pruning new API
