+++ test/pruning/test_pruning.py
@@ -1,5 +1,3 @@
-import os
-import shutil
 import unittest
 
 import torch
@@ -7,7 +5,7 @@ import torchvision
 import torch.nn as nn
 from neural_compressor.data import Datasets
 from neural_compressor.experimental.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
-from neural_compressor.pruning import Pruning, WeightPruningConfig
+from neural_compressor.pruner.pruning import Pruning, WeightPruningConfig
 
 
 class TestPruning(unittest.TestCase):
