+++ test/pruning/test_pruning_schedulers.py
@@ -1,5 +1,3 @@
-import os
-import shutil
 import unittest
 
 import torch
@@ -8,7 +6,7 @@ import torch.nn as nn
 
 from neural_compressor.data import Datasets
 from neural_compressor.experimental.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
-from neural_compressor.pruning import Pruning, WeightPruningConfig
+from neural_compressor.pruner.pruning import Pruning, WeightPruningConfig
 
 local_schedulers_config = [
     {
